Adds actions on collision (#1)
* added onCollision function & refactor

Triggers on collision with another entity or block
refactored the findCollisionsWithData to be more elegant

* Improves projectile behavior and collisions

Improves projectile spawning by offsetting its initial position based on the player's direction and adds initial speed.

Implements projectile deletion on collision with other entities.

Adds `doReactCollision` to Collisionable to allow entities to ignore collision physics, sets it to false for projectiles, and ensures collisions are only handled if both entities involved react to collisions.

Adds a system for deleting entities to prevent concurrent modification exceptions and manage entity removal in a thread-safe manner.
This includes adding a `toBeDeletedEntities` list to the `EntitiesManager`
and processing it at the end of each server tick.

Cleans up entities managers.

* changed packet constructor argument to whole entity

* Small refactor & started collision exceptions

* Added collision exceptions.